[Quant] Consolidate GPTQ quantization: remove gptq.py, rename gptq_marlin.py to gptq.py
The Marlin kernels support all GPTQ use cases (4-bit and 8-bit symmetric),
making the legacy exllama-based gptq.py redundant. This consolidation:

- Renames gptq_marlin.py → gptq.py; GPTQMarlinConfig.get_name() now returns "gptq"
- Removes override_quantization_method() from GPTQMarlinConfig (no longer needed
  since gptq IS gptq_marlin now — the override loop falls through naturally)
- Deprecates "gptq_marlin" as a quantization method; auto-converts to "gptq" in
  _verify_quantization() so --quantization gptq_marlin still works with a warning
- Updates moe_wna16, inc, openpangu_vl, gptq_utils to import from gptq instead
  of gptq_marlin; simplifies gptq_utils.override_config() to remove dead branch
- Removes GPTQLinearMethod (exllama kernel) from linear.py WEIGHT_LOADER_V2_SUPPORTED
- Updates tests to use GPTQMarlinLinearMethod; removes symFalse test case since
  asymmetric GPTQ is not supported by the Marlin kernels
